Rename conditions modules
(PR #4501)
