+++ Lib/fontbakery/checks/fontval.py
@@ -3,7 +3,7 @@ import tempfile
 
 from fontbakery.prelude import check, ERROR, FAIL, INFO, PASS, WARN, Message
 from fontbakery.utils import exit_with_install_instructions
-from fontbakery.profiles.shared_conditions import (
+from fontbakery.shared_conditions import (
     is_cff,
     is_variable_font,
 )  # pylint: disable=unused-import
