+++ Lib/fontbakery/checks/fontwerk.py
@@ -4,7 +4,7 @@ Checks for Fontwerk <https://fontwerk.com/>
 
 from fontbakery.prelude import check, PASS, FAIL, Message
 from fontbakery.constants import FsSelection, MacStyle
-from fontbakery.profiles.shared_conditions import (  # pylint: disable=unused-import
+from fontbakery.shared_conditions import (  # pylint: disable=unused-import
     is_variable_font,
     has_wght_axis,
 )
@@ -174,7 +174,7 @@ def com_fontwerk_check_inconsistencies_between_fvar_stat(ttFont):
 )
 def com_fontwerk_check_style_linking(ttFont):
     """Checking style linking entries"""
-    from fontbakery.profiles.shared_conditions import is_italic, is_bold
+    from fontbakery.shared_conditions import is_italic, is_bold
 
     errs = []
     if is_bold(ttFont):
