+++ Lib/fontbakery/checks/googlefonts/conditions.py
@@ -17,7 +17,7 @@ from fontbakery.utils import exit_with_install_instructions
 # used to inform get_module_profile whether and how to create a profile
 from fontbakery.fonts_profile import profile_factory  # noqa:F401 pylint:disable=W0611
 
-from .shared_conditions import style
+from fontbakery.profiles.shared_conditions import style
 
 
 GF_TAGS_SHEET_URL = (
