+++ Lib/fontbakery/checks/loca.py
@@ -2,7 +2,7 @@ from fontbakery.callable import check
 from fontbakery.status import PASS, FAIL
 from fontbakery.message import Message
 
-from fontbakery.profiles.shared_conditions import (  # pylint: disable=unused-import
+from fontbakery.shared_conditions import (  # pylint: disable=unused-import
     is_ttf,
 )
