+++ Lib/fontbakery/checks/name.py
@@ -9,7 +9,7 @@ from fontbakery.constants import (
 )
 from fontbakery.utils import markdown_table
 
-from fontbakery.profiles.shared_conditions import (  # pylint: disable=unused-import
+from fontbakery.shared_conditions import (  # pylint: disable=unused-import
     is_ttf,
     glyph_metrics_stats,
     style,
