+++ Lib/fontbakery/checks/stat.py
@@ -5,7 +5,7 @@ from fontbakery.message import Message
 from fontbakery.status import FAIL, PASS, WARN, SKIP
 from fontbakery.utils import bullet_list
 
-from fontbakery.profiles.shared_conditions import (  # pylint: disable=unused-import
+from fontbakery.shared_conditions import (  # pylint: disable=unused-import
     is_variable_font,
     has_STAT_table,
     style,
