+++ Lib/fontbakery/shared_conditions.py
